third_party/WebKit/LayoutTests/compositing/overflow/invisible-descendants-should-not-affect-opt-in-expected.txt

Iteration 0: Passed
Iteration 0, layer tree
Iteration 1: Passed
Iteration 1, no layer tree
Iteration 2: Passed
Iteration 2, layer tree
